fix(stackblitz): drag and drop examples not working
* Since the `DragDropModule` is not added to the StackBlitz template, the drag & drop examples don't work if opened in StackBlitz.
* Moves the `DemoMaterialModule` into a separate file because having one big files is confusing and makes it hard to debug the StackBlitz.
